### Importar Residentes para PR

Tela destinada à importação de residentes para o programa PR, permitindo o cadastro de múltiplos residentes de forma rápida e padronizada por meio de um arquivo de importação.

O usuário poderá selecionar o arquivo contendo os dados dos residentes, realizar o carregamento e validar as informações antes de concluir a importação. O sistema deverá apresentar mensagens de sucesso, alerta ou erro, indicando possíveis inconsistências encontradas durante o processamento.

A tela deve permitir:

* Selecionar e enviar o arquivo de importação;
* Validar os dados antes da inclusão;
* Identificar registros inválidos ou duplicados;
* Exibir o resultado da importação;
* Informar a quantidade de registros importados com sucesso;
* Apresentar os registros que não puderam ser importados e o respectivo motivo;
* Permitir a correção dos dados e uma nova tentativa de importação.

+++ backend/src/repositories/indicatorManagementRepository.js
@@ -261,8 +261,8 @@ export async function deleteRecord(id, userId) {
 export async function recordsForCalculation(centerId, year) {
   const { rows } = await query(
     `SELECT * FROM indicator_records WHERE innovation_center_id=$1 AND deleted_at IS NULL
-       AND ((year IS NULL AND event_at IS NULL AND start_date IS NULL) OR year=$2 OR EXTRACT(YEAR FROM event_at)=$2
-         OR (start_date<=make_date($2,12,31) AND (end_date IS NULL OR end_date>=make_date($2,1,1))))`,
+       AND ((year IS NULL AND event_at IS NULL AND start_date IS NULL) OR year=$2::int OR EXTRACT(YEAR FROM event_at)=$2::int
+         OR (start_date<=make_date($2::int,12,31) AND (end_date IS NULL OR end_date>=make_date($2::int,1,1))))`,
     [centerId, year],
   );
   return rows;
